[TASK] Restructure code
