+++ Scripts/CommonScript.php
@@ -6,88 +6,10 @@ namespace Ochorocho\Tdk\Scripts;
 
 use Composer\Script\Event;
 use Composer\Util\ProcessExecutor;
-use Symfony\Component\Filesystem\Exception\IOException;
 use Symfony\Component\Filesystem\Filesystem;
 
-class InitializeScript extends BaseScript
+class CommonScript extends BaseScript
 {
-    public static function enableHooks(Event $event)
-    {
-        $questions = [
-            [
-                'method' => 'enableCommitMessageHook',
-                'message' => 'Setup Commit Message Hook? [<fg=cyan;options=bold>y</>/n] ',
-                'default' => true
-            ],
-            [
-                'method' => 'enablePreCommitHook',
-                'message' => 'Setup Pre Commit Hook? [<fg=cyan;options=bold>y</>/n] ',
-                'default' => true
-            ],
-        ];
-
-        $force = (bool)(GitScript::getArguments($event->getArguments())['force'] ?? getenv('TDK_HOOK_FORCE_CREATE') ?? false);
-        foreach ($questions as $question) {
-            if ($force) {
-                $answer = true;
-            } else {
-                $answer = $event->getIO()->askConfirmation($question['message'], $question['default']);
-            }
-
-            if ($answer) {
-                $method = $question['method'];
-                static::$method($event);
-            }
-        }
-    }
-
-    public static function removeHooks(Event $event)
-    {
-        $filesystem = new Filesystem();
-        $filesystem->remove([
-            self::$coreDevFolder . '/.git/hooks/pre-commit',
-            self::$coreDevFolder . '/.git/hooks/commit-msg',
-        ]);
-    }
-
-    private static function enableCommitMessageHook(Event $event)
-    {
-        $filesystem = new Filesystem();
-
-        try {
-            $targetCommitMsg = self::$coreDevFolder . '/.git/hooks/commit-msg';
-            $filesystem->copy(self::$coreDevFolder . '/Build/git-hooks/commit-msg', $targetCommitMsg);
-
-            if (!is_executable($targetCommitMsg)) {
-                $filesystem->chmod($targetCommitMsg, 0755);
-            }
-
-            $event->getIO()->write('<info>Created Commit Message Hook</info>');
-        } catch (IOException $e) {
-            $event->getIO()->writeError('<warning>Exception:enableCommitMessageHook:' . $e->getMessage() . '</warning>');
-        }
-    }
-
-    private static function enablePreCommitHook(Event $event)
-    {
-        if (DIRECTORY_SEPARATOR === '\\') {
-            return;
-        }
-        $filesystem = new Filesystem();
-        try {
-            $targetPreCommit = self::$coreDevFolder . '/.git/hooks/pre-commit';
-            $filesystem->copy(self::$coreDevFolder . '/Build/git-hooks/unix+mac/pre-commit', $targetPreCommit);
-
-            if (!is_executable($targetPreCommit)) {
-                $filesystem->chmod($targetPreCommit, 0755);
-            }
-
-            $event->getIO()->write('<info>Created Pre Commit Hook</info>');
-        } catch (IOException $e) {
-            $event->getIO()->writeError('<warning>Exception:enablePreCommitHook:' . $e->getMessage() . '</warning>');
-        }
-    }
-
     public static function createDdevConfig(Event $event)
     {
         // Only ask for ddev config if ddev command is available
@@ -95,9 +17,9 @@ class InitializeScript extends BaseScript
         $test = $windows ? 'where' : 'command -v';
 
         if (is_executable(trim(shell_exec($test . ' ddev') ?? ''))) {
-            $ddevProjectName = GitScript::getArguments($event->getArguments())['project-name'] ?? getenv('TDK_CREATE_DDEV_PROJECT_NAME') ?? false;
+            $ddevProjectName = self::getArguments($event->getArguments())['project-name'] ?? getenv('TDK_CREATE_DDEV_PROJECT_NAME') ?? false;
             if (!$ddevProjectName) {
-                $skip = isset(GitScript::getArguments($event->getArguments())['no']) ?? false;
+                $skip = isset(self::getArguments($event->getArguments())['no']) ?? false;
                 if ($skip) {
                     $createConfig = false;
                 } else {
@@ -111,7 +33,7 @@ class InitializeScript extends BaseScript
                 }
             }
 
-            $validator = self::validateDdevProjectName();
+            $validator = ValidatorScript::projectName();
 
             if (!$ddevProjectName) {
                 $defaultProjectName = basename(getcwd());
@@ -152,7 +74,7 @@ class InitializeScript extends BaseScript
             'var',
         ];
 
-        $force = GitScript::getArguments($event->getArguments())['force'] ?? false;
+        $force = self::getArguments($event->getArguments())['force'] ?? false;
 
         if ($force) {
             $answer = true;
@@ -165,32 +87,6 @@ class InitializeScript extends BaseScript
             $filesystem->remove($filesToDelete);
             $event->getIO()->write('<info>Done deleting files.</info>');
         }
-    }
-
-    public static function showSummary(Event $event): void
-    {
-        $coreFolder = self::$coreDevFolder;
-        $summary = <<<EOF
-
-💡For more Details read the docs:
-* Setting up Gerrit (ssh):
-  https://docs.typo3.org/m/typo3/guide-contributionworkflow/master/en-us/Account/GerritAccount.html
-* Git Setup:
-  https://docs.typo3.org/m/typo3/guide-contributionworkflow/master/en-us/Setup/Git/Index.html
-* Setup your IDE:
-  https://docs.typo3.org/m/typo3/guide-contributionworkflow/master/en-us/Setup/SetupIde.html
-* runTests.sh docs still apply, but don't forget to cd into '$coreFolder':
-  https://docs.typo3.org/m/typo3/guide-contributionworkflow/master/en-us/Testing/Index.html
-
-<fg=yellow;options=bold>To be able to push to Gerrit, you need to add your public key, see https://review.typo3.org/settings/#SSHKeys</>
-EOF;
-
-        $event->getIO()->write($summary);
-    }
-
-    public static function done(Event $event): void
-    {
-        $event->getIO()->write('<info>🎉 Happy days ... TYPO3 Composer CoreDev Setup done!</info>');
     }
 
     public static function doctor(Event $event): void
