+++ composer.json
@@ -14,7 +14,7 @@
         "post-root-package-install": "@tdk:setup",
         "post-create-project-cmd": [
             "@tdk:help",
-            "Ochorocho\\Tdk\\Scripts\\InitializeScript::done"
+            "Ochorocho\\Tdk\\Scripts\\MessageScript::done"
         ],
         "tdk:setup": [
             "@tdk:clone",
@@ -24,12 +24,12 @@
             "@tdk:set-commit-template"
         ],
         "tdk:clone": "Ochorocho\\Tdk\\Scripts\\GitScript::cloneRepository",
-        "tdk:clear": "Ochorocho\\Tdk\\Scripts\\InitializeScript::removeFilesAndFolders",
-        "tdk:doctor": "Ochorocho\\Tdk\\Scripts\\InitializeScript::doctor",
-        "tdk:remove-hooks": "Ochorocho\\Tdk\\Scripts\\InitializeScript::removeHooks",
-        "tdk:enable-hooks": "Ochorocho\\Tdk\\Scripts\\InitializeScript::enableHooks",
-        "tdk:ddev-config": "Ochorocho\\Tdk\\Scripts\\InitializeScript::createDdevConfig",
-        "tdk:help": "Ochorocho\\Tdk\\Scripts\\InitializeScript::showSummary",
+        "tdk:clear": "Ochorocho\\Tdk\\Scripts\\CommonScript::removeFilesAndFolders",
+        "tdk:doctor": "Ochorocho\\Tdk\\Scripts\\CommonScript::doctor",
+        "tdk:remove-hooks": "Ochorocho\\Tdk\\Scripts\\HookScript::remove",
+        "tdk:enable-hooks": "Ochorocho\\Tdk\\Scripts\\HookScript::enable",
+        "tdk:ddev-config": "Ochorocho\\Tdk\\Scripts\\CommonScript::createDdevConfig",
+        "tdk:help": "Ochorocho\\Tdk\\Scripts\\MessageScript::summary",
         "tdk:set-commit-template": "Ochorocho\\Tdk\\Scripts\\GitScript::setCommitTemplate",
         "tdk:set-git-config": "Ochorocho\\Tdk\\Scripts\\GitScript::setGitConfig",
         "tdk:apply-patch": "Ochorocho\\Tdk\\Scripts\\GitScript::applyPatch",
